+++ demo/src/main/java/com/example/ecapp/dto/OrderItemResponse.java
@@ -16,8 +16,8 @@ public class OrderItemResponse {
   private Long productId;
   private String productName;
   private String productDescription;
-  private int price;
+  private long price;
   private int quantity;
-  private int subtotal; // price * quantity
+  private long subtotal; // price * quantity
   
 }
